restful-booker herokuapp page how to dark mode
https://restful-booker.herokuapp.com/apidoc/index.html#api-Booking-UpdateBooking

First, download Stylus as Google Browser and after downloading, open it and paste the codes by pressing the write new style button. Translate the button that is (everything) as (URL). Paste the URLY above to where you pasted the codes and save it, then you can reach the dark mode page by searching for that URL on Google.

Öncelikle Stylus'u Google Browser olarak indirin ve indirdikten sonra açıp yeni stil yaz butonuna basarak kodları yapıştırın. (her şey) olan butonu (URL) olarak çevirin. Yukarıdaki URLY'yi kodları yapıştırdığınız yere yapıştırın ve kaydedin, ardından Google'da o URL'yi arayarak karanlık mod sayfasına ulaşabilirsiniz.